// src/shared/messages/error-messages.enum.ts
export enum ErrorMessages {
    REGISTER_DRONE_FAILURE = 'Drone registration failed',
    DRONE_NOT_FOUND = 'Drone not found',
    DRONE_ALREADY_EXISTS = 'Drone with this serial number already exists',
    DRONE_LOAD_FAILURE = 'Unable to load Drone',
    MEDICATION_NOT_EMPTY = 'Each medication must not be empty',
    MAX_WEIGHT_EXCEEDED = 'Max weight exceeded',
    LOW_BATTERY = 'Low battery: Cannot load medications',
    INVALID_MEDICATION_CODE = 'Invalid medication code',
    INVALID_MEDICATION_NAME = 'Invalid medication name',
}