// src/shared/messages/success-messages.enum.ts
export enum SuccessMessages {
  WELCOME_HOME = 'Welcome Home',
  REGISTER_DRONE_SUCCESS = 'Drone registered successfully',
  LOAD_DRONE_SUCCESS = 'Drone loaded successfully',
  MEDICATIONS_RETRIEVED = 'Medications retrieved successfully',
  DRONE_FETCHED = 'Drones fetched successfully',
  DRONE_FETCHED_DETAILS = 'Drone details fetched successfully',
}